refactor default avatar and default menu/meal cover

+++ grubcat/fanju/apis.py
@@ -2,6 +2,7 @@
 from api_auth import UserObjectsOnlyAuthorization
 from datetime import datetime, timedelta, date
 from django.conf.urls.defaults import url
+from django.conf import settings
 from django.contrib import auth
 from django.contrib.auth import logout
 from django.db.utils import IntegrityError
@@ -464,7 +465,8 @@ class UserResource(EOResource):
             filename = contentFile.name
             user_to_query.cropping = "" #cropping is not supported by app yet so clear it
             user_to_query.avatar.save(filename, contentFile)
-            if os.path.exists(old_avatar_path) and user_to_query.avatar.path != old_avatar_path:
+            if os.path.exists(
+                    old_avatar_path) and user_to_query.avatar.path != old_avatar_path and settings.DEFAULT_FEMALE_AVATAR not in old_avatar_path and settings.DEFAULT_MALE_AVATAR not in old_avatar_path:
                 os.remove(old_avatar_path)
             # xmpp_client.syncProfile(user_to_query)
             user_resource = UserResource()
